% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{create_discrete_scale}
\alias{create_discrete_scale}
\title{Create a function that returns a discrete scale}
\usage{
create_discrete_scale(aes = c("colour", "fill", "linetype"), pal = NULL)
}
\arguments{
\item{aes}{A character aesthetic. Allowed values are "colour", "fill" and
"linetype" (partial matching).}

\item{pal}{A character vector containing the palette.}
}
\value{
A function that returns a discrete scale when called.
}
\description{
Create a discrete scale using nish colour scheme.
}
