% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{create_gradient2_scale}
\alias{create_gradient2_scale}
\title{Create a function that returns a gradient2 scale}
\usage{
create_gradient2_scale(type = NULL, aes = c("colour", "fill"))
}
\arguments{
\item{type}{A character denoting the gradient2 type. Allowed values are "pn1"
and "pn2".}

\item{aes}{A character aesthetic. Allowed values are "colour" and "fill".}
}
\value{
A function that returns a diverging colour gradient scale when called.
}
\description{
Create a gradient2 scale (diverging colour gradient) using nish colour scheme.
}
\details{
If options ggplot2.continuous.colour and ggplot2.continuous.fill are unset,
gradient is used instead of continuous.
}
\examples{
scale_colour_gradient2_nish_pos_neg <- nish:::create_gradient2_scale(
type = "pn1", aes = "colour")

}
