% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{create_gradient_scale}
\alias{create_gradient_scale}
\title{Create a function that returns a gradient scale}
\usage{
create_gradient_scale(
  col = c("blue", "grey", "pink"),
  aes = c("colour", "fill")
)
}
\arguments{
\item{col}{A character string of gradient nuanse. Allowed values are "blue",
"grey" and "pink".}

\item{aes}{A character aesthetic. Allowed values are "colour" and "fill".}
}
\value{
A function that returns a two colour gradient scale when called.
}
\description{
Create a gradient scale using nish colour scheme.
}
\details{
If options ggplot2.continuous.colour and ggplot2.continuous.fill are unset,
gradient is used instead of continuous.
}
\examples{
scale_fill_gradient_nish_blue <- nish:::create_gradient_scale(col = "blue",
aes = "fill")

}
