% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{disc}
\alias{disc}
\alias{scale_colour_nish_blue}
\alias{scale_color_nish_blue}
\alias{scale_fill_nish_blue}
\alias{scale_colour_nish_grey}
\alias{scale_color_nish_grey}
\alias{scale_fill_nish_grey}
\alias{scale_colour_nish_pink}
\alias{scale_color_nish_pink}
\alias{scale_fill_nish_pink}
\alias{scale_colour_nish_mixed}
\alias{scale_color_nish_mixed}
\alias{scale_fill_nish_mixed}
\alias{scale_colour_nish_pos_neg}
\alias{scale_color_nish_pos_neg}
\alias{scale_fill_nish_pos_neg}
\alias{scale_colour_nish_pos_neg2}
\alias{scale_color_nish_pos_neg2}
\alias{scale_fill_nish_pos_neg2}
\alias{scale_linetype_nish}
\title{Discrete nish scales}
\usage{
scale_colour_nish_blue(...)

scale_color_nish_blue(...)

scale_fill_nish_blue(...)

scale_colour_nish_grey(...)

scale_color_nish_grey(...)

scale_fill_nish_grey(...)

scale_colour_nish_pink(...)

scale_color_nish_pink(...)

scale_fill_nish_pink(...)

scale_colour_nish_mixed(...)

scale_color_nish_mixed(...)

scale_fill_nish_mixed(...)

scale_colour_nish_pos_neg(...)

scale_color_nish_pos_neg(...)

scale_fill_nish_pos_neg(...)

scale_colour_nish_pos_neg2(...)

scale_color_nish_pos_neg2(...)

scale_fill_nish_pos_neg2(...)

scale_linetype_nish(...)
}
\arguments{
\item{...}{Additional arguments are passed to
\code{\link[ggplot2]{discrete_scale}}().}
}
\description{
Discrete nish scales
}
