% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{ggplot_add.conditional_theme}
\alias{ggplot_add.conditional_theme}
\title{Change theme settings conditional on 'geom_*' used in the plot}
\usage{
\method{ggplot_add}{conditional_theme}(object, plot, object_name)
}
\arguments{
\item{object}{An object (theme) to add to the plot.}

\item{plot}{The ggplot object to add \code{object} to.}

\item{object_name}{The name of the object to add.}
}
\value{
A ggplot object.
}
\description{
Method to change theme settings conditional on 'geom_*' used in the plot
}
\details{
The original idea was to use:
\code{plot$theme <- ggplot2:::add_theme(plot$theme, object, "object")}
for this purpose but \code{ggplot2:::add_theme}() is an internal ggplot2
function and thus not exported.

Instead, the current theme in \code{object} consist of \code{theme_nish_*}
settings. These theme properties are further modified based on 'geom_*' in
this method.

\code{plot$theme} may initially (as input to this method) be an empty list if
no \code{theme}() calls have been performed or set in working environment.
\code{object} always replaces an existing theme in \code{plot$theme}.

Note: The conditional part of the theme settings does not work with
\code{ggplot2::theme_set}(). Explicit \code{theme_nish_*()} calls required to
enable conditional theme settings.
}
