% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{grad}
\alias{grad}
\alias{scale_colour_gradient_nish_blue}
\alias{scale_color_gradient_nish_blue}
\alias{scale_fill_gradient_nish_blue}
\alias{scale_colour_gradient_nish_grey}
\alias{scale_color_gradient_nish_grey}
\alias{scale_fill_gradient_nish_grey}
\alias{scale_colour_gradient_nish_pink}
\alias{scale_color_gradient_nish_pink}
\alias{scale_fill_gradient_nish_pink}
\title{Gradient nish scales}
\usage{
scale_colour_gradient_nish_blue(...)

scale_color_gradient_nish_blue(...)

scale_fill_gradient_nish_blue(...)

scale_colour_gradient_nish_grey(...)

scale_color_gradient_nish_grey(...)

scale_fill_gradient_nish_grey(...)

scale_colour_gradient_nish_pink(...)

scale_color_gradient_nish_pink(...)

scale_fill_gradient_nish_pink(...)
}
\arguments{
\item{...}{Additional arguments are passed to
\code{\link[ggplot2:scale_gradient]{scale_colour_gradient}}() or
\code{\link[ggplot2:scale_gradient]{scale_fill_gradient}}().}
}
\description{
Gradient nish scales
}
