% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{grad2}
\alias{grad2}
\alias{scale_colour_gradient2_nish_pos_neg}
\alias{scale_color_gradient2_nish_pos_neg}
\alias{scale_fill_gradient2_nish_pos_neg}
\alias{scale_colour_gradient2_nish_pos_neg2}
\alias{scale_color_gradient2_nish_pos_neg2}
\alias{scale_fill_gradient2_nish_pos_neg2}
\title{Gradient2 nish scales}
\usage{
scale_colour_gradient2_nish_pos_neg(...)

scale_color_gradient2_nish_pos_neg(...)

scale_fill_gradient2_nish_pos_neg(...)

scale_colour_gradient2_nish_pos_neg2(...)

scale_color_gradient2_nish_pos_neg2(...)

scale_fill_gradient2_nish_pos_neg2(...)
}
\arguments{
\item{...}{Additional arguments are passed to
\code{ggplot2::scale_colour_gradient2}() or
\code{ggplot2::scale_fill_gradient2}().}
}
\description{
Gradient2 nish scales
}
