% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{gradn}
\alias{gradn}
\alias{scale_colour_gradientn_nish}
\alias{scale_color_gradientn_nish}
\alias{scale_fill_gradientn_nish}
\title{Gradientn nish scales}
\usage{
scale_colour_gradientn_nish(...)

scale_color_gradientn_nish(...)

scale_fill_gradientn_nish(...)
}
\arguments{
\item{...}{Additional arguments are passed to
\code{\link[ggplot2:scale_gradient]{scale_colour_gradientn}}() or
\code{\link[ggplot2:scale_gradient]{scale_fill_gradientn}}(). For example,
use argument 'colours' to provide your own vector of colours for the
n-colour gradient.}
}
\description{
Gradientn nish scales
}
