% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\docType{data}
\name{nish_colours}
\alias{nish_colours}
\title{nish (main) colour palette that includes the majority of the colours used}
\format{
An object of class \code{character} of length 20.
}
\usage{
nish_colours
}
\description{
nish (main) colour palette that includes the majority of the colours used
}
\examples{
nish_colours
}
\keyword{datasets}
