% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{theme_nish}
\alias{theme_nish}
\alias{theme_nish_blue}
\alias{theme_nish_pink}
\alias{theme_nish_white}
\title{nish themes}
\usage{
theme_nish_blue()

theme_nish_pink()

theme_nish_white()
}
\description{
nish themes with either light blue, light pink or white background. All nish
themes are based on the same underlying internal base theme.
}
\details{
Allows for conditional theme settings, see
\code{\link[nish:theme_cond]{nish::ggplot_add.conditional_theme}}().
}
\note{
\itemize{
\item Conditional theme settings does not work with
\code{ggplot2::theme_set}(). For example, by using
\code{old_theme <- theme_set(theme_nish_blue())}.
However, \code{theme_nish_*()} can be be called explicitly to enable
conditional theme settings.
\item axis.ticks are set which does not fully agree with requirements.
}

To adjust nish themes (for developers):
\itemize{
\item Modify theme settings in \code{theme_nish_base}() for settings
applicable to all \code{theme_nish_*()} colour themes.
\item Adjust \code{theme_nish_blue}(), \code{theme_nish_pink}() or
\code{theme_nish_white}() for settings that are specific to each colour
theme.
\item Adjust \code{nish::ggplot_add.conditional_theme}() for any
conditional theme logic.}
}
