% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nish.R
\name{theme_nish}
\alias{theme_nish}
\alias{theme_nish_blue}
\alias{theme_nish_pink}
\alias{theme_nish_white}
\title{nish themes}
\usage{
theme_nish_blue()

theme_nish_pink()

theme_nish_white()
}
\description{
nish themes with light blue ("#e4f2ff"), light pink ("#fcebe3") and white
backgrounds.
}
